Add Eslite scraper file

+++ kingstone.py
@@ -10,7 +10,6 @@ import requests
 import time
 
 
-
 client = MongoClient('localhost', 27017)
 db = client.Bookstores
 
@@ -360,35 +359,37 @@ if __name__=='__main__':
     daily_change_tracker(catalog_today, catalog_yesterday, 'kingstone_pid', new_prodcut_catalog, unfound_product_catalog)
 
 
-    # Step 5: Reading catalog and scraped single product info 
+    # Step 5: Reading [new_prodcut_catalog] of today and scraped single product info 
     product_catalog = new_prodcut_catalog.find({'track_date': TODAY})
     product_list = convert_mongo_object_to_list(product_catalog)
-    multi_scrapers(
-        worker_num = 20, 
-        list_to_scrape = product_list, 
-        url_to_scrape = PRODUCT_PAGE, 
-        target_id_key = 'kingstone_pid', 
-        db_to_insert = product_info, 
-        scraper_func = get_product_info, 
-        insert_func = mongo_insert,
-        slicing=True
-    )
+    if len(product_list) > 0:
+        multi_scrapers(
+            worker_num = 20, 
+            list_to_scrape = product_list, 
+            url_to_scrape = PRODUCT_PAGE, 
+            target_id_key = 'kingstone_pid', 
+            db_to_insert = product_info, 
+            scraper_func = get_product_info, 
+            insert_func = mongo_insert,
+            slicing=True
+        )
 
     # Step 6: Reading [unfound_product_catalog], add current back to [catalog_today], phased out to [phase_out_product_catalog]
     #         Delete after finishing scraping
     product_catalog = unfound_product_catalog.find()
     product_list = convert_mongo_object_to_list(product_catalog)
-    multi_scrapers(
-        worker_num = 20, 
-        list_to_scrape = product_list, 
-        url_to_scrape = PRODUCT_PAGE, 
-        target_id_key = 'kingstone_pid', 
-        db_to_insert = catalog_today, 
-        scraper_func = phased_out_checker, 
-        insert_func = mongo_insert,
-        slicing=True
-    )
-    db.drop_collection(unfound_product_catalog)
+    if len(product_list) > 0:
+        multi_scrapers(
+            worker_num = 20, 
+            list_to_scrape = product_list, 
+            url_to_scrape = PRODUCT_PAGE, 
+            target_id_key = 'kingstone_pid', 
+            db_to_insert = catalog_today, 
+            scraper_func = phased_out_checker, 
+            insert_func = mongo_insert,
+            slicing=True
+        )
+        db.drop_collection(unfound_product_catalog)
 
     
     # Step 7. Delete catalog of 7 days age, EX: today is '2021-10-26', so delete '2021-10-19'
